# docs/swagger.yaml
definitions:
...
  model.CopyRequest:
    properties:
      from_paths:
        description: 源路径列表
        items:
          type: string
        minItems: 1
        type: array
      to_path:
        description: 目标路径
        type: string
    required:
    - from_paths
    - to_path
    type: object
  model.DeleteRequest:
    properties:
      paths:
        description: 要删除的路径列表
        items:
          type: string
        minItems: 1
        type: array
    required:
    - paths
    type: object
...
  model.MkdirRequest:
    properties:
      path:
        description: 目录路径
        type: string
    required:
    - path
    type: object
  model.MoveRequest:
    properties:
      from_paths:
        description: 源路径列表
        items:
          type: string
        minItems: 1
        type: array
      to_path:
        description: 目标路径
        type: string
    required:
    - from_paths
    - to_path
    type: object
  model.RecycleDeleteRequest:
    properties:
      fsids:
        description: 要删除的文件ID列表
        items:
          type: integer
        minItems: 1
        type: array
    required:
    - fsids
    type: object
  model.RecycleRestoreRequest:
    properties:
      fsids:
        description: 要恢复的文件ID列表
        items:
          type: integer
        minItems: 1
        type: array
    required:
    - fsids
    type: object
  model.Response:
    properties:
      code:
        description: 状态码：0-成功，非0-失败
        type: integer
      data:
        description: 响应数据
      message:
        description: 响应消息
        type: string
    type: object
...
  model.ShareCancelRequest:
    properties:
      share_ids:
        description: 要取消的分享ID列表
        items:
          type: integer
        minItems: 1
        type: array
    required:
    - share_ids
    type: object
  model.ShareSetRequest:
    properties:
      password:
        description: 提取码（可选）
        type: string
      paths:
        description: 要分享的路径列表
        items:
          type: string
        minItems: 1
        type: array
      period:
        description: 有效期（天数，0表示永久）
        type: integer
    required:
    - paths
    type: object
  model.TransferRequest:
    properties:
      collect:
        description: 是否归档到同名目录
        type: boolean
      download:
        description: 转存后是否自动下载
        type: boolean
      extract_code:
        description: 提取码（可选）
        type: string
      share_url:
        description: 分享链接
        type: string
    required:
    - share_url
    type: object
...
info:
  contact: {}
paths:
...
  /api/cp:
    post:
      consumes:
      - application/json
      description: 复制文件或目录
      parameters:
      - description: 复制请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.CopyRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: 复制文件
      tags:
      - 文件管理
...
  /api/mkdir:
    post:
      consumes:
      - application/json
      description: 在网盘中创建新目录
      parameters:
      - description: 创建目录请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.MkdirRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: 创建目录
      tags:
      - 文件管理
  /api/mv:
    post:
      consumes:
      - application/json
      description: 移动文件或重命名文件
      parameters:
      - description: 移动/重命名请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.MoveRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: 移动/重命名文件
      tags:
      - 文件管理
  /api/pwd:
    get:
      consumes:
      - application/json
      description: 获取当前用户的工作目录路径
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/model.Response'
      summary: 获取当前工作目录
      tags:
      - 工作目录
  /api/recycle/clear:
    post:
      consumes:
      - application/json
      description: 清空回收站中所有文件
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: 清空回收站
      tags:
      - 回收站
  /api/recycle/delete:
    post:
      consumes:
      - application/json
      description: 彻底删除回收站中的文件（不可恢复）
      parameters:
      - description: 删除请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.RecycleDeleteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: 彻底删除
      tags:
      - 回收站
  /api/recycle/list:
    get:
      consumes:
      - application/json
      description: 列出回收站中的文件
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: 列出回收站
      tags:
      - 回收站
  /api/recycle/restore:
    post:
      consumes:
      - application/json
      description: 恢复回收站中的文件
      parameters:
      - description: 恢复请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.RecycleRestoreRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: 恢复文件
      tags:
      - 回收站
  /api/rm:
    post:
      consumes:
      - application/json
      description: 删除网盘中的文件或目录
      parameters:
      - description: 删除请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.DeleteRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: 删除文件
      tags:
      - 文件管理
...
  /api/share/cancel:
    post:
      consumes:
      - application/json
      description: 取消指定的分享链接
      parameters:
      - description: 取消分享请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.ShareCancelRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: 取消分享
      tags:
      - 分享管理
  /api/share/list:
    get:
      consumes:
      - application/json
      description: 列出当前账户的所有分享
      parameters:
      - description: 页码
        in: query
        name: page
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: 列出分享
      tags:
      - 分享管理
  /api/share/set:
    post:
      consumes:
      - application/json
      description: 创建文件的分享链接
      parameters:
      - description: 创建分享请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.ShareSetRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: 创建分享
      tags:
      - 分享管理
  /api/transfer:
    post:
      consumes:
      - application/json
      description: 将他人的分享链接转存到自己网盘
      parameters:
      - description: 转存请求
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/model.TransferRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.Response'
      summary: 转存分享链接
      tags:
      - 转存管理
...
swagger: "2.0"